Updates:
1. The Humble Bundle site itself is now only loaded one time to determine the
name of the bundle. After that, Humble's bundle statistics provider (pubnub) is
called instead using that bundle name. This is a much more lightweight call
(just returns a JSON object with the bundle statistics), and pubnub expects this
to be called repeatedly as that's the behavior by simply by browing to the
Humble Bundle page and leaving it open.
2. Scheduled task to delete Humble Bundle log files has been updated so that
both the bundle name file and the price log are deleted
3. Since our check is now so much more efficient, I have changed the scheduled
task to run only on Tuesdays starting at 1:00, but run every 20 seconds
4. The code has been cleaned up a good amount. I'm using Invoke-Webrequest now
rather than the hacked-together readHTML function that was used previously.
I'm also spoofing the Internet Explorer user agent to make it appear more human
5. A bug regarding a price comparison with the first price check has been
fixed.

Work to be done:
* Remove the need of a scheduled task to delete the humble bundle files. This
should be done directly by the script at the time the new Bundle is introduced
  * This may be as simple as doing a check at 1:00 for the bundle name and
  re-initializing if it is different than what is saved in the file
* Determine if Humble Bundle adheres to daylight savings time, and make sure
that the scheduled task(s) are configured correctly

--HG--
branch : feature/usepubnub